Add URL and Minor Fix to FibSeq()
Modified the URL to the problem so that instead of going to the Project Euler homepage, it instead goes to the page of the problem itself.
I also reduced the redundancy of the i++ code in FebSeq()'s while loop. Instead of having it at the end of both if and else statements, I simply added a single i++ at the end of the loop.