Add Solutions for Problem 3
problem3.c is the first attempt at the problem and problem3Alternate.c is a refined method that works a bit quicker. The former increments through all numbers (less than the target), whereas the latter uses the tree method for factoring and reduces the amount of computation by reassigning the target number as a factor when a factor is found. Using the fact that a factors of a number's factor are also factors of that original number, the number of iterations is drastically reduced as factors are found.